:recycle: :fire: Remove create() and associated types and methods from register_delegate
