+++ packages/lisk-transactions/test/index.ts
@@ -23,12 +23,6 @@ describe('transaction', () => {
 				.and.be.a('function');
 		});
 
-		it('should have #registerDelegate', () => {
-			return expect(transaction)
-				.to.have.property('registerDelegate')
-				.and.be.a('function');
-		});
-
 		it('should have #castVotes', () => {
 			return expect(transaction)
 				.to.have.property('castVotes')
